0.0.13b1

+++ simpla/SettingsAdmin.php
@@ -80,7 +80,7 @@ class SettingsAdmin extends Simpla
                 $clear_image_cache = true;
             }
             if ($this->config->images['crop_factor'] != $this->request->post('crop_factor')) {
-                $this->config->images['crop_factor'] = (int)$this->request->post('crop_factor');
+                $this->config->images['crop_factor'] = (float)$this->request->post('crop_factor');
                 $clear_image_cache = true;
             }
             if ($this->config->images['bg_color'] != $this->request->post('bg_color')) {
